+++ NEWS.md
@@ -1,3 +1,5 @@
 # fluxnet (development version)
 
+* Added `flux_extract()` for extracting data from downloaded .zip files.
 * Added functions for downloading FLUXNET data
+